Add command line and lambdas

+++ clock.cpp
@@ -1,11 +1,21 @@
 //==============================================================================
-// clock.cpp
+// clock.cpp		Clock/calendar with Google calendar using gtkmm
+//					Nigel Hewitt 2022
+//
+//					Feel free to cut and stick while usual disclaimers apply.
+//					I had trouble finding a straight forward gtkmm example so,
+//					now I have gathered bits from all over the web, it only
+//					seems fair that I publish my own.
+//					Tested on Pi3/4 and WSL Debian
 //==============================================================================
 //
-// sudo apt install g++ make
-// sudo apt install libgtk-3-dev libgtkmm-3.0-dev
-
-// Compile this with the attached makefile
+// spaced with tab=4
+//
+// Requires:
+//		sudo apt install g++ make gdb
+//		sudo apt install libgtk-3-dev libgtkmm-3.0-dev
+//
+// Compile with the attached makefile to get the includes and the libraries
 //		make
 //
 // Pi-Clock aka 'Pi in the Sky' is set to execute at startup
@@ -17,18 +27,19 @@
 // 2022-03-27  use local time not UTC
 // 2022-07-27  add dow and date
 // 2022-10-09  add the Google Calendar stuff
-// 2022-10-12  switch to gtkmm
+// 2022-10-12  switch to gtkmm and full on C++
 // 2022-10-14  finish and fine tune as a C++ example
 // 2022-10-17  add css compile error reporting
-// 2022-10-26  add code for Z duration (new today)
-
+// 2022-10-26  add code for Z duration (seen new today)
+// 2022-10-27  add command arguments
+//
 // For Eclipse this requires the pkg-config plugin
 //   Help | Eclipse Market place
 //   put pkg-config in the search box | Go
 //     Install and restart
 // Add to project:
 //   Project | Properties | C++ Build | Settings
-//     Pkg-config (initially an off screen to the right tab)
+//     Pkg-config (initially an off screen tab to the right)
 //     check gtk-3.0 and gtkmm-3.0  | Apply and continue
 //
 //==============================================================================
@@ -44,7 +55,8 @@
 #include <iostream>
 
 // Define some CSS so we can set colours and fonts and stuff
-// I break it into lines with \n so we get nice error messages
+// I break it into lines with \n so we get useful error messages
+// if it fails to compile
 
 static const char *css =
 "window {\n"							// top level Gtk::Window
@@ -78,47 +90,49 @@ static const char *css =
 ;
 
 // Now the class that defines our main window
+// I have coded it with the functions 'inline' C# style
 
-class Clock : public Gtk::Window {
+class CLOCK : public Gtk::Window {
 protected:
+	// As there only one instance of CLOCK we can cheat the static/dynamic vibe
+	inline static CLOCK* self;
+	
 	// Member widgets:
 	Gtk::Fixed fixed;				// a widget container with fixed coordinates
 	Gtk::Button close, refresh;		// buttons
 	Gtk::Label time, day, date;		// blocks of text
 	Gtk::Label slot[5];				// more text for the calendar entries
 
-	// Signal handlers:  these get called on a button click et al
-	void on_close_clicked(){
-		Gtk::Window::close();
-	}
-	void on_refresh_clicked(){
-		Ticks = 12;
-	}
-
+	bool bTest{false};				// used when testing
+	
 public:
-	Clock(){	// the constructor for the window
+	CLOCK() = delete;							// no default constructor
+	CLOCK(Glib::RefPtr<Gtk::Application> app){	// the constructor for the window
+		self = this;
 		set_title("Pi-Clock");
 		set_border_width(10);
 
 		// Select the 'fixed' container so we get absolute coordinates.
-		// Beware: the basic Gtk::Window can only contain one widget so this
-		// is that, but this is a container for lots of widgets
+		// Beware: the basic Gtk::Window can only contain one widget but this
+		// is a widget that contains lots of widgets
 		// My old screen is a 1440 x 900 - old but that's why it's on the clock
-		fixed.set_size_request(1440-30, 900-52);  // minimum (preferred) so screen size
-												  //	less borders, toolbar and title et al
+		fixed.set_size_request(1440-30, 900-52);  // minimum (preferred) so
+												  // screen size less borders,
+												  // toolbar and title et al
 		add(fixed);			// put the Gtk::Fixed in the Gtk::Window
 
-		// Arrange for some CSS to do colours and fonts
+		// Arrange for the CSS to do colours and fonts
 		auto context = get_style_context();
 		auto provider = Gtk::CssProvider::create();
 		try{
 			provider->load_from_data(css);
 		}
 		catch(const Gtk::CssProviderError& e){
-			// If there is a syntax error the gtkmm code eats the error message.
-			// So, if it fails, recompile it with a more basic function that
-			// actually outputs a useful error description with line/column.
-			// It will still throw but at least you get useful stuff to read
+			// If there is a syntax error the gtkmm code helpfully eats the
+			// error message. So, if it fails, recompile it with a more basal
+			// function that actually throws with a useful error description
+			// with line and column numbers.
+			// It will still throw but at least you get help to fix it
 			if(e.code()==Gtk::CssProviderError::SYNTAX){
 				GtkCssProvider *provider = gtk_css_provider_new();
 				gtk_css_provider_load_from_data(provider, css, -1, nullptr);
@@ -130,17 +144,28 @@ public:
 							provider, GTK_STYLE_PROVIDER_PRIORITY_USER);
 
 		// Give the labels CSS names so we can distinguish them
-		time.set_name("aval");			// ie: label#aval
+		time.set_name("aval");			// ie: use label#aval
 		day.set_name("bval");
 		date.set_name("bval");
 		for(int i=0; i<5; ++i)
 			slot[i].set_name("sval1");
 
-		// Connect the buttons to their service routines
-		close.signal_clicked().connect(sigc::mem_fun(*this,
-													&Clock::on_close_clicked));
-		refresh.signal_clicked().connect(sigc::mem_fun(*this,
-													&Clock::on_refresh_clicked));
+		// Connect the buttons to their service routines as lambdas
+		close.signal_clicked().connect([this]{ return Gtk::Window::close(); });
+		refresh.signal_clicked().connect([this]{ Ticks = 12; });
+												
+		// And the command line argument receiver
+		// more messy as it is a static
+		app->signal_command_line().connect(
+			[app](const Glib::RefPtr<Gio::ApplicationCommandLine>& command_line){
+				int argc = 0;
+				char** argv = command_line->get_arguments(argc);
+				self->do_command(argc, argv);	// call a member function of CLOCK
+				
+				app->activate();	// beware: this is in the default
+									// on_command() so we must do it too
+				return 0;
+	     	}, false);
 
 		// Set the buttons' texts and put them into the fixed container
 		close.set_label("Close");
@@ -158,17 +183,29 @@ public:
 		// The final step is to display all these newly created widgets...
 		show_all_children();
 
-		// Make a ticker to call tick() every 1000mS
-		Glib::signal_timeout().connect(sigc::bind_return(sigc::mem_fun(*this,
-											&Clock::tick), true), 1000);
+		// Make a timer to call CLOCK::tick() every 1000mS
+		// I'll use a lambda again to save a layer of indirection
+		Glib::signal_timeout().connect([this]() { return this->tick(); }, 1000);
 	}
-	virtual ~Clock(){}
-
+	virtual ~CLOCK(){}		// default clean-ups only
+	
+	// receive the command args
+	void do_command(int argc, char* argv[])
+	{
+		for(int i=0; i<argc; ++i){
+			if(strcmp(argv[i], "-t")==0)
+				bTest = true;
+		}
+	}
+	
 	// Ticker at 1 per second
-	int Ticks{20};			// delay the first fetch for ten seconds
-	char today[12]{};		// used to colour the lines for today
+	int Ticks{25};			// delay the first fetch for fifteen seconds
+	int Retries{0};			// limit the fast retries
+	char today[12]{};		// used to colour the lines for 'today'
 
-	// update the time, day and date
+	// Update the time, day and date
+	int oldDOW{9};			// used to trigger the refresh of day oriented
+	
 	void setDisplay()
 	{
 		char temp[30];
@@ -179,19 +216,22 @@ public:
 		sprintf(temp, "%02d:%02d:%02d", t->tm_hour, t->tm_min, t->tm_sec);
 		time.set_text(temp);
 
-		const char* dow[] = { "Sunday",   "Monday", "Tuesday", "Wednesday",
-							  "Thursday", "Friday", "Saturday"  };
-		day.set_text(dow[t->tm_wday]);
+		// the rest only changes if the day changes
+		if(t->tm_wday != oldDOW){
+			oldDOW = t->tm_wday;
+			const char* dow[] = { "Sunday",   "Monday", "Tuesday", "Wednesday",
+								  "Thursday", "Friday", "Saturday"  };
+			day.set_text(dow[t->tm_wday]);
 
-		sprintf(temp, "%02d-%02d-%04d", t->tm_mday, t->tm_mon+1, 1900+t->tm_year);
-		date.set_text(temp);
+			sprintf(temp, "%02d-%02d-%04d", t->tm_mday, t->tm_mon+1, 1900+t->tm_year);
+			date.set_text(temp);
 
-		// Since we have the time here it seems messy to redo it somewhere else
-		// to make a value to compare to the Google calendar stuff:
-		sprintf(today, "%04d-%02d-%02d", 1900+t->tm_year, t->tm_mon+1, t->tm_mday);
+			// Make a value to compare to the Google calendar stuff:
+			sprintf(today, "%04d-%02d-%02d", 1900+t->tm_year, t->tm_mon+1, t->tm_mday);
+		}
 	}
 	
-	// update the calendar
+	// Update the calendar
 	void setCalendar()
 	{
 		// The events file has four sorts of entries, all day, timed and errors
@@ -199,25 +239,26 @@ public:
 		// 2022-10-13T12:00:00+01:00 Lunch with Robin\n
 		// 2022-11-01T21:00:00Z Recycling  (first seen 26/10/2022)
 		// * something bad happened\n
+		// its stderr output are sent to response.edc so we can try
+		// and fail responsibly
 		
 #define CALDIR	"/home/pi/calendar"
 		const char* eventsFile   = CALDIR "/events.txt";
 		const char* responseFile = CALDIR "/response.edc";
 
-		if(--Ticks==10){		// at 10 seconds to go trigger the calendar reader
-			if(fork()==0){		// go multi-threaded
+		if(--Ticks==10 && !bTest){	// at 10 seconds to go run the calendar
+			if(fork()==0){			// go multi-threaded
 				chdir(CALDIR);
 				unlink(responseFile);
 				unlink(eventsFile);
 				char command[100];
 				sprintf(command, "python clock.py 2> %s", responseFile);
-		std::cout <<  command << std::endl;
 				system(command);
 				exit(0);		// kill the forked thread
 			}
 		}
 		if(Ticks<=0){
-			Ticks = 60*60;		// reset for one hour
+			Ticks = bTest ? 60 : 60*60;		// reset for one hour
 			int i=0;
 			FILE* f = fopen(eventsFile, "r");
 			if(f){
@@ -244,7 +285,7 @@ public:
 							// copy the start time but discard the duration
 							while(j<19) text2[k++] = text1[j++];
 							text2[k++] = ' ';
-							// is it a '+' time or a Z time?
+							// is it a '+' time or a 'Z' time?
 							if(text1[j]=='+') j += 7;
 							else              j += 2;
 						}
@@ -261,9 +302,14 @@ public:
 						slot[i].set_text(text2);
 					}
 				}
+				Retries = 0;
 				fclose(f);
 			}
-			else{			// if the events file failed to open
+			else{				// if the events file failed to open
+				// If it fails a couple of times retry but if it's stuck revert
+				// to the one hour schedule.
+				if(++Retries<4)
+					Ticks = 60*2;	// give it two minutes and then try again
 				FILE* f2 = fopen(responseFile, "r");
 				if(f2){
 					char buffer[200];
@@ -285,7 +331,7 @@ public:
 				}
 			}
 			if(i==0){						// response file failed too
-				slot[i].set_name("sval1");		// red
+				slot[i].set_name("sval1");	// red
 				slot[i++].set_text("** Data failed to fetch **");					
 			}
 			for( ; i<5; ++i){			// blank the rest of the display
@@ -294,17 +340,28 @@ public:
 			}
 		}
 	}
-	void tick()
+	bool tick()
 	{
 		setDisplay();
 		setCalendar();
+		return true;
 	}
 };
 
+
 int main(int argc, char *argv[])
 {
-	auto app = Gtk::Application::create(argc, argv, "clock.app");
+	// Command line arguments are a pain under gtkmm so I will try to clean it.
+	// We add the APPLICATION_HANDLES_COMMAND_LINE flag so we get sent the args
+	// then we hook up a receiver callback in CLOCK to handle them
+	// This way gtkmm gets a first look at the args and act on and takes out
+	// those that belong to it and then passes the rest down to us.
+	
+	auto app = Gtk::Application::create(argc, argv, "clock.app",
+							Gio::APPLICATION_HANDLES_COMMAND_LINE);
+	
+	CLOCK Clock(app);
+	
 	// Show the window and returns when it is closed.
-	Clock clock;
-	return app->run(clock);
+	return app->run(Clock);
 }
